PRSD-1284: Add subject line to email template tests
