+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/emailModels/EmailTemplateModelsTests.kt
@@ -7,6 +7,7 @@ import org.springframework.test.util.ReflectionTestUtils
 import uk.gov.communities.prsdb.webapp.constants.enums.RegistrationNumberType
 import uk.gov.communities.prsdb.webapp.database.entity.LocalAuthority
 import uk.gov.communities.prsdb.webapp.models.dataModels.RegistrationNumberDataModel
+import uk.gov.communities.prsdb.webapp.testHelpers.EmailTemplateMetadata
 import java.net.URI
 
 class EmailTemplateModelsTests {
@@ -111,11 +112,13 @@ class EmailTemplateModelsTests {
     @MethodSource("templateList")
     fun `EmailTemplateModels hashmaps have keys that match the parameters in their markdown templates`(testData: EmailTemplateTestData) {
         // Arrange
-        var storedBody = javaClass.getResource(testData.markdownLocation)?.readText() ?: ""
-        var parameters = getParametersFromBody(storedBody).distinct()
+        val storedBody = javaClass.getResource(testData.markdownLocation)?.readText() ?: ""
+        val storedMetadata = EmailTemplateMetadata.metadataList.single { metadata -> metadata.id == testData.model.templateId.idValue }
+
+        val parameters = getParametersFromBody(storedMetadata.subject + "\n\n" + storedBody).distinct()
 
         // Act
-        var modelHashMap = testData.model.toHashMap()
+        val modelHashMap = testData.model.toHashMap()
 
         // Assert
         Assertions.assertEquals(parameters.size, modelHashMap.size)
